Revert "added types.d.ts"
