+++ src/redux/types.d.ts
@@ -1,0 +1,11 @@
+// export const ALLOCATE_EMPLOYEE_ASSET = 'ALLOCATE_EMPLOYEE_ASSET'
+// declare interface SetAllocateAsset {
+//   type: typeof ALLOCATE_EMPLOYEE_ASSET
+//   payload: {
+//     message: string
+//   }
+// }
+
+declare interface MyInterface {
+  id: string
+}
